+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 tarantool thread via syncronization primitives (channels) and low-level cbus api.
 - `tarantool::time::Instant` a custom implementation of std-like `Instant` with more saturating operations
 and support of the `fiber_clock` API.
+- `r#async::sleep` - an async friendly analog of `fiber::sleep`.
 
 ### Fixed
 - `log::Log::enabled` implementation for TarantoolLogger no longer ignores the